+++ src/components/cloud/EditCloudButton.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { Button } from '@/components/ui/button';
 import { Modal } from '@/components/shared/Modal';
 import { CloudForm } from '@/components/cloud/cloudForm/CloudForm';
-import { CloudFormData, cloudFormSchema } from '@/lib/cloud-form-schema';
+import { CloudFormData, cloudFormSchema } from '@/lib/cloudFormSchema';
 import { Cloud } from '@/types';
 
 interface EditCloudButtonProps {
