+++ src/components/cloud/cloudForm/ScanFrequencyFields.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Control, useWatch } from 'react-hook-form';
-import { CloudFormData } from '@/lib/cloud-form-schema';
+import { CloudFormData } from '@/lib/cloudFormSchema';
 
 import {
   FormControl,
